Added wrapper around `Executor` to require static reference.
- Polling the Executor is safe if the Executor is static. However,
  adding a static lifetime to self when polling makes the borrow checker
  unhappy. By adding a wrapper struct for the Executor that holds a
  static mutable reference, we can poll safely from the wrapper.
